requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-jet==1.0.8
django-rest-swagger==2.2.0
djangorestframework==3.9.4
google-api-python-client==1.4.1
httplib2==0.13.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2client==4.1.3
openapi-codec==1.3.2
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
